+++ examples/extending/components/fa-star.vue
@@ -1,0 +1,11 @@
+<script type="text/javascript">
+import {
+    fa_star
+} from '../../../glyphs'
+import FaBaseGlyph from '../../../src/raters/font-awesome/FaBaseGlyph.vue';
+export default FaBaseGlyph.extend({
+    created() {
+        this.points = [fa_star];
+    }
+});
+</script>