+++ src/index.js
@@ -3,8 +3,8 @@ import HeartRating from './raters/hearts/heart-rating.vue';
 import FaRating from './raters/font-awesome/font-awesome-rating.vue';
 import ImageRating from './raters/images/image-rating.vue';
 import BaseRating from './raters/BaseRating.vue';
+import RateIt from './raters/rate-it.vue';
 import mixins from './mixins/index.js';
-import Glyphs from './raters/font-awesome/glyphs.js';
 import FaBaseGlyph from './raters/font-awesome/FaBaseGlyph.vue';
 
 const raters = {
@@ -24,6 +24,6 @@ export {
     FaRating,
     ImageRating,
     BaseRating,
-    Glyphs,
-    FaBaseGlyph
+    FaBaseGlyph,
+    RateIt
 }