+++ changelog.md
@@ -2,8 +2,11 @@
 - Project follows [Semantic Versioning](http://semver.org/)
 
 ## [Unreleased]
-- Added export of individual font awesome glyphs
-- Added extendable base glyph for font-awesome to allow use of specific glyph
+- Added export of individual font awesome glyphs.
+- Added extendable base glyph for font-awesome to allow use of glyps without importing all of them.
+- Added new rate-it component for passing custom rating items.
+- Moved glyphs file to root so it can be accessed by `import {fa_glyph} from `vue-rate-it/glyphs;` 
+- Added new "extendable" example.
 
 ## Version 1.0 (2017-04-26)
 - Initial release